requirements.txt
discord.py
odmantic
dnspython
